feat: implement pygments-block
Use `<pre><code>` to wrap the block of code. This is the proper way to
annotate blocks of code (https://stackoverflow.com/questions/4611591/code-vs-pre-vs-samp-for-inline-and-block-code-snippets).

Use div-wrapper for the block of code, so that the code is more
composable and I can use things like `display: flex` on that div in CSS.